+++ hdfs-connector/src/test/java/com/oracle/bmc/hdfs/TestDeprecatedBmcFileSystemContract.java
@@ -7,7 +7,9 @@ package com.oracle.bmc.hdfs;
 
 import java.util.HashSet;
 import java.util.Set;
+import java.util.concurrent.TimeUnit;
 
+import com.oracle.bmc.hdfs.contract.BmcContract;
 import com.oracle.bmc.hdfs.contract.deprecated.DeprecatedBmcContract;
 import com.oracle.bmc.hdfs.store.BmcDataStore;
 import org.apache.hadoop.conf.Configuration;
@@ -15,6 +17,7 @@ import org.apache.hadoop.fs.FileStatus;
 import org.apache.hadoop.fs.FileSystemContractBaseTest;
 import org.apache.hadoop.fs.Path;
 import org.junit.Before;
+import org.junit.BeforeClass;
 import org.junit.experimental.categories.Category;
 
 import static org.junit.Assert.assertEquals;
@@ -23,8 +26,19 @@ import static org.junit.Assert.assertTrue;
 
 @Category({IntegrationTestCategory.class})
 public class TestDeprecatedBmcFileSystemContract extends FileSystemContractBaseTest {
+
+    public static Path INITIAL_WORKING_DIRECTORY;
+
+    @BeforeClass
+    public static void setUpStatic() throws Exception {
+        final Configuration configuration = new Configuration();
+        final DeprecatedBmcContract contract = new DeprecatedBmcContract(configuration);
+        contract.init();
+        INITIAL_WORKING_DIRECTORY = contract.getTestFileSystem().getWorkingDirectory();
+    }
+
     @Before
-    protected void setUp() throws Exception {
+    public void setUp() throws Exception {
         final Configuration configuration =
                 new Configuration() {
                     @Override
@@ -36,6 +50,9 @@ public class TestDeprecatedBmcFileSystemContract extends FileSystemContractBaseT
         final DeprecatedBmcContract contract = new DeprecatedBmcContract(configuration);
         contract.init();
         super.fs = contract.getTestFileSystem();
+
+        // reset the working directory to avoid test-to-test influence
+        fs.setWorkingDirectory(INITIAL_WORKING_DIRECTORY);
     }
 
     @Override
@@ -43,6 +60,11 @@ public class TestDeprecatedBmcFileSystemContract extends FileSystemContractBaseT
         super.tearDown();
 
         super.fs.delete(new Path("/existingobjects"), true);
+    }
+
+    @Override
+    protected int getGlobalTimeout() {
+        return (int) TimeUnit.SECONDS.toMillis(60);
     }
 
     @Override