+++ hdfs-connector/src/test/java/com/oracle/bmc/hdfs/contract/inmemory/TestInMemoryBmcContractSeek.java
@@ -1,0 +1,13 @@
+package com.oracle.bmc.hdfs.contract.inmemory;
+
+import com.oracle.bmc.hdfs.contract.BmcContract;
+import com.oracle.bmc.hdfs.contract.TestBmcContractSeek;
+import org.apache.hadoop.conf.Configuration;
+import org.apache.hadoop.fs.contract.AbstractFSContract;
+
+public class TestInMemoryBmcContractSeek extends TestBmcContractSeek {
+    @Override
+    protected AbstractFSContract createContract(final Configuration conf) {
+        return new BmcContract.InMemory(conf);
+    }
+}